feat: introduce `wasmcloud-compat` crate
