+++ Cargo.toml
@@ -124,6 +124,7 @@ wasi-common = { version = "11", default-features = false }
 wasm-metadata = { version = "0.10.1", default-features = false }
 wasmcloud-actor = { path = "./crates/actor" }
 wasmcloud-actor-derive = { path = "./crates/actor/derive" }
+wasmcloud-compat = { path = "./crates/compat" }
 wasmcloud-component-adapters = { git = "https://github.com/wasmCloud/wasmcloud-component-adapters" }
 wasmcloud-control-interface = { version = "0.28", default-features = false }
 wasmcloud-host = { path = "./crates/host" }
