+++ crates/compat/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod http;
+pub mod keyvalue;
+pub mod messaging;
+
+pub use http::{Request as HttpRequest, Response as HttpResponse};
