+++ crates/compat/src/messaging.rs
@@ -1,10 +1,4 @@
-use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
-use wasmcloud_provider_sdk::{
-    core::{LinkDefinition, WasmCloudEntity},
-    error::ProviderInvocationError,
-    Context,
-};
 
 /// A message to be published
 #[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
@@ -68,62 +62,4 @@ pub struct SubMessage {
     #[serde(with = "serde_bytes")]
     #[serde(default)]
     pub body: Vec<u8>,
-}
-
-pub struct Handler<'a> {
-    ld: &'a LinkDefinition,
-}
-
-impl<'a> Handler<'a> {
-    pub fn new(ld: &'a LinkDefinition) -> Self {
-        Self { ld }
-    }
-
-    pub async fn handle_message(&self, msg: SubMessage) -> Result<(), ProviderInvocationError> {
-        let connection = wasmcloud_provider_sdk::provider_main::get_connection();
-
-        let client = connection.get_rpc_client();
-        let origin = WasmCloudEntity {
-            public_key: self.ld.provider_id.clone(),
-            link_name: self.ld.link_name.clone(),
-            contract_id: "wasmcloud:messaging".to_string(),
-        };
-        let target = WasmCloudEntity {
-            public_key: self.ld.actor_id.clone(),
-            ..Default::default()
-        };
-
-        let data = wasmcloud_provider_sdk::serialize(&msg)?;
-
-        let response = client
-            .send(origin, target, "MessageSubscriber.HandleMessage", data)
-            .await?;
-
-        if let Some(e) = response.error {
-            Err(ProviderInvocationError::Provider(e))
-        } else {
-            Ok(())
-        }
-    }
-}
-
-/// The Messaging interface describes a service
-/// that can deliver messages
-/// wasmbus.contractId: wasmcloud:messaging
-/// wasmbus.providerReceive
-#[async_trait]
-pub trait Messaging {
-    /// returns the capability contract id for this interface
-    fn contract_id() -> &'static str {
-        "wasmcloud:messaging"
-    }
-    /// Publish - send a message
-    /// The function returns after the message has been sent.
-    /// If the sender expects to receive an asynchronous reply,
-    /// the replyTo field should be filled with the
-    /// subject for the response.
-    async fn publish(&self, ctx: Context, arg: PubMessage) -> Result<(), String>;
-    /// Request - send a message in a request/reply pattern,
-    /// waiting for a response.
-    async fn request(&self, ctx: Context, arg: RequestMessage) -> Result<ReplyMessage, String>;
 }
