+++ crates/host/Cargo.toml
@@ -42,6 +42,7 @@ ulid = { workspace = true, features = ["std"] }
 url = { workspace = true, features = ["serde"] }
 uuid = { workspace = true, features = ["serde"] }
 wascap = { workspace = true }
+wasmcloud-compat = { workspace = true }
 wasmcloud-control-interface = { workspace = true }
 wasmcloud-runtime = { workspace = true }
