+++ crates/providers/Cargo.toml
@@ -36,4 +36,5 @@ tracing-futures = { version = "0.2", default-features = false }
 tracing-opentelemetry = { version = "0.20", default-features = false }
 warp = { version = "0.3", default-features = false }
 wascap = { version = "0.8", default-features = false }
+wasmcloud-compat = { path = "../compat", default-features = false }
 wasmcloud-provider-sdk = { path = "../provider-sdk", default-features = false }
