+++ crates/providers/http-server/Cargo.toml
@@ -28,6 +28,7 @@ tokio = { workspace = true, features = ["full"] }
 toml = { workspace = true, features = ["parse"] }
 tracing = { workspace = true }
 warp = { workspace = true, features = ["tls"] }
+wasmcloud-compat = { workspace = true }
 wasmcloud-provider-sdk = { workspace = true, features = ["otel"] }
 
 [lib]
