+++ crates/providers/http-server/src/lib.rs
@@ -32,16 +32,12 @@
 //! by the all of the server green threads.
 //!
 
-// TODO: These types should be defined via WIT
-pub mod wasmcloud_interface_httpserver;
-
 mod hashmap_ci;
 mod settings;
 
 pub use settings::{load_settings, ServiceSettings, CONTENT_LEN_LIMIT, DEFAULT_MAX_CONTENT_LEN};
-pub use wasmcloud_interface_httpserver::{
-    HttpRequest as Request, HttpResponse as Response, HttpServer as Server,
-};
+
+pub use wasmcloud_compat::http::{Request, Response};
 
 pub(crate) use hashmap_ci::make_case_insensitive;
 
@@ -55,11 +51,13 @@ use bytes::Bytes;
 use flume::{bounded, Receiver, Sender};
 use futures::Future;
 use http::header::HeaderMap;
-
 use tokio::task::JoinHandle;
 use tracing::{debug, error, info, trace, Instrument};
-use warp::{filters::cors::Builder, path::FullPath, Filter};
-use wasmcloud_provider_sdk::{core::LinkDefinition, error::ProviderInvocationError};
+use warp::filters::cors::Builder;
+use warp::path::FullPath;
+use warp::Filter;
+use wasmcloud_provider_sdk::core::LinkDefinition;
+use wasmcloud_provider_sdk::error::ProviderInvocationError;
 
 /// errors generated by this crate
 #[derive(Debug, thiserror::Error)]
