+++ crates/providers/kv-redis/Cargo.toml
@@ -21,4 +21,5 @@ serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 tracing = { workspace = true }
+wasmcloud-compat = { workspace = true }
 wasmcloud-provider-sdk = { workspace = true, features = ["otel"] }
