--- crates/providers/kv-redis/src/lib.rs
@@ -1,2 +1,0 @@
-// TODO: These types should be defined via WIT
-pub mod wasmcloud_interface_keyvalue;
