+++ crates/providers/nats/src/lib.rs
@@ -1,3 +1,1 @@
 pub mod otel;
-// TODO: These types should be defined via WIT
-pub mod wasmcloud_interface_messaging;
